Move files to package
