+++ src/main/java/net/dv8tion/jda/api/interactions/commands/Command.java
@@ -22,6 +22,7 @@ import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.ISnowflake;
 import net.dv8tion.jda.api.interactions.DiscordLocale;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
+import net.dv8tion.jda.api.interactions.commands.localization.LocalizationMap;
 import net.dv8tion.jda.api.interactions.commands.privileges.CommandPrivilege;
 import net.dv8tion.jda.api.requests.RestAction;
 import net.dv8tion.jda.api.requests.restaction.CommandEditAction;
@@ -39,7 +40,7 @@ import java.time.OffsetDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static net.dv8tion.jda.api.interactions.commands.LocalizationMap.UNMODIFIABLE_CHECK;
+import static net.dv8tion.jda.api.interactions.commands.localization.LocalizationMap.UNMODIFIABLE_CHECK;
 
 /**
  * Represents a Discord slash-command.
